moves mocked test from integration to unit directory
