+++ tests/integration/test_bot.py
@@ -12,7 +12,7 @@ from nightshift.comms.storage import Drive, get_credentials
 
 
 @pytest.mark.local
-def test_logger_on_run(env_var, caplog, test_data):
+def test_logger_on_run(env_var, caplog, test_data_rich):
 
     with caplog.at_level(logging.INFO):
         run(env="prod")
